requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.14.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.23.0
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.7.0.84
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu11==2.19.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
open-clip-torch==2.24.0
opencv-python==4.9.0.80
packaging==24.0
pillow==10.3.0
platformdirs==4.2.1
protobuf==4.25.3
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
sentry-sdk==2.1.1
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.12
termcolor==2.4.0
threadpoolctl==3.5.0
timm==0.9.16
tokenizers==0.19.1
torch==2.2.2+cu118
torch_geometric==2.5.3
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.4
transformers==4.40.2
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
wandb==0.17.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.18.1
